pyproject.toml
[project]
name = "discord-joinerbot"
version = "0.1.0"
description = "A discord bot to send notifications when people are in voice channels."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
]
